requirements.txt
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-text-splitters>=0.3.0
langchain-community>=0.3.0
chromadb>=0.5.0
langchain-chroma>=0.2.0
tree-sitter>=0.23.0
tree-sitter-java>=0.23.0
tree-sitter-javascript>=0.23.0
PyGithub>=2.3.0
atlassian-python-api>=3.41.0
Jinja2>=3.1.0
typer[all]>=0.12.0
python-dotenv>=1.0.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
rich>=13.7.0
tiktoken>=0.7.0
